uwierzytelnianie

+++ backend/app/db/advert_interface.py
@@ -2,7 +2,7 @@
 import psycopg2
 from fastapi import APIRouter, Query, Path, Depends, Body, Form, HTTPException
 from typing import Optional, List
-from ..models import Advert
+from ..models import Advert, User
 from .user_interface import DatabaseDetails, get_connection
 
 db = DatabaseDetails()
@@ -80,21 +80,33 @@ class DBGetAdvert:
 class DBEditAdvert:
 
     # @get_connection
-    def add_advert(advert: Advert) -> None:
+    def add_advert(advert: Advert, current_user: User) -> None:
         conn = psycopg2.connect(conn_string)
         cursor = conn.cursor()
 
+        author_id = current_user.user_id
+
         cursor.execute("INSERT INTO adverts (latitude, longitude, date, price, author_id, description, title, images) VALUES (%s, %s, %s, %s, %s, %s, %s, %s);", 
-        (advert.latitude, advert.longitude, advert.date, advert.price, advert.author_id, advert.description, advert.title, advert.images))
+        (advert.latitude, advert.longitude, advert.date, advert.price, author_id, advert.description, advert.title, advert.images))
 
         # conn.commit()
         cursor.close()
         conn.close()
     
     # @get_connection
-    def delete_advert(advert_id: int) -> None:
+    def delete_advert(advert_id: int, current_user: User) -> None:
         conn = psycopg2.connect(conn_string)
         cursor = conn.cursor()
+
+        cursor.execute("SELECT * FROM adverts WHERE advert_id = %s;", (advert_id,))
+        rows = cursor.fetchall()
+        if len(rows) == 0:
+            raise HTTPException(status_code=404, detail="Advert not found")
+        author_id = rows[0][4]
+
+        if current_user.user_id != author_id:
+            raise HTTPException(status_code=403, detail="You are not the author of this advert")
+
 
         cursor.execute("DELETE FROM adverts WHERE advert_id = %s;", (advert_id,))
 
@@ -102,13 +114,23 @@ class DBEditAdvert:
         cursor.close()
         conn.close()
 
-    @get_connection
-    def update_advert(advert_id: int, advert: Advert) -> None:
+    # @get_connection
+    def update_advert(advert_id: int, advert: Advert, current_user: User) -> None:
         conn = psycopg2.connect(conn_string)
         cursor = conn.cursor()
 
+        cursor.execute("SELECT * FROM adverts WHERE advert_id = %s;", (advert_id,))
+        rows = cursor.fetchall()
+        if len(rows) == 0:
+            raise HTTPException(status_code=404, detail="Advert not found")
+        author_id = rows[0][4]
+
+        if current_user.user_id != author_id:
+            raise HTTPException(status_code=403, detail="You are not the author of this advert")
+
+
         cursor.execute("UPDATE adverts SET latitude = %s, longitude = %s, date = %s, price = %s, author_id = %s, description = %s, title = %s, images = %s WHERE advert_id = %s;", 
-        (advert.latitude, advert.longitude, advert.date, advert.price, advert.author_id, advert.description, advert.title, advert.images, advert_id))
+        (advert.latitude, advert.longitude, advert.date, advert.price, author_id, advert.description, advert.title, advert.images, advert_id))
 
         # conn.commit()
         cursor.close()
